+++ launches/USLI_Launch_Beta/ArduinoTimer.h
@@ -2,21 +2,12 @@
 #define ARDUINO_TIMER
 #include <chrono>
 
-auto start_chrono_time = std::chrono::high_resolution_clock::now();
+extern std::chrono::time_point<std::chrono::high_resolution_clock> start_chrono_time;
 
-double seconds() {
-    auto current_chrono_time = std::chrono::high_resolution_clock::now();
-    return std::chrono::duration_cast<std::chrono::seconds>(current_chrono_time - start_chrono_time).count();
-}
+double seconds();
 
-size_t millis() {
-    auto current_chrono_time = std::chrono::high_resolution_clock::now();
-    return std::chrono::duration_cast<std::chrono::milliseconds>(current_chrono_time - start_chrono_time).count();
-}
+size_t millis();
 
-size_t micros() {
-    auto current_chrono_time = std::chrono::high_resolution_clock::now();
-    return std::chrono::duration_cast<std::chrono::microseconds>(current_chrono_time - start_chrono_time).count();
-}
+size_t micros();
 
 #endif
