+++ launches/USLI_Launch_Beta/FlightMath.h
@@ -18,7 +18,6 @@
 #include <math.h>
 #include <stdlib.h>
 
-const double GRAVITY = 9.80665;
 /**
  * Function that returns net value of three values. 
  * Useful to detect any motion
@@ -28,10 +27,7 @@ const double GRAVITY = 9.80665;
  * @param z third value
  * @return net value of three vales
 */
-double net_value(float x, float y, float z)
-{
-    return sqrt(x*x + y*y + z*z);
-}
+double net_value(float x, float y, float z);
 
 /**
  * Function that calculates vertical velocity from altitude change. 
@@ -42,11 +38,7 @@ double net_value(float x, float y, float z)
  * @param elapsed_time Elapsed time in seconds between current and previous altitude reads
  * @return vertical_velocity
 */
-double vertical_velocity(float altitude, float prev_altitude, double elapsed_time)
-{
-    double delta_alt = altitude - prev_altitude;
-    return (double)(delta_alt / elapsed_time);
-}
+double vertical_velocity(float altitude, float prev_altitude, double elapsed_time);
 
 /**
  * Function that calculates projected altitude from current altitude, velocity and acceleration. 
@@ -56,11 +48,7 @@ double vertical_velocity(float altitude, float prev_altitude, double elapsed_tim
  * @param acceleration Previous elapsed time of data read
  * @return projected altitude
 */
-double projected_altitude(double altitude, double velocity, double acceleration)
-{
-    return fabs(velocity * velocity) / (2.0 * fabs(acceleration + GRAVITY)) * 
-        log(fabs(acceleration) / GRAVITY) + altitude;
-}
+double projected_altitude(double altitude, double velocity, double acceleration);
 
 /**
  * Function that calculates altitude from pressure. 
@@ -68,22 +56,10 @@ double projected_altitude(double altitude, double velocity, double acceleration)
  * @param pressure Current pressure read
  * @return altitude
 */
-double altitude_from_pressure(double pressure)
-{
-    double seaLevelhPa = 1013.25;
-    double alt;
+double altitude_from_pressure(double pressure);
 
-    pressure /= 100;
-    alt = 44330 * (1.0 - pow(pressure / seaLevelhPa, 0.1903));
-
-    return alt;
-}
-
-double low_pass_filter(double previous_value, double current_value, double elapsed_time) {
-    const double RC = 0.3;
-    const double alpha = elapsed_time / (RC + elapsed_time);
-    return (alpha * current_value) + (1.0 - alpha) * previous_value;
-}
+/* TODO Add Docs */
+double low_pass_filter(double previous_value, double current_value, double elapsed_time);
 
 /**
  * Function that calculates projected altitude from current altitude and velocity. 
@@ -93,9 +69,6 @@ double low_pass_filter(double previous_value, double current_value, double elaps
  * @param velocity Vertical velocity
  * @return projected altitude
 */
-double simple_projected_altitude(double altitude, double velocity)
-{
-    return altitude + (velocity * velocity) / (2.0 * GRAVITY);
-}
+double simple_projected_altitude(double altitude, double velocity);
 
 #endif
