+++ launches/USLI_Launch_Beta/flight.cpp
@@ -1,7 +1,15 @@
+#include "flight.h"
 
 #include "FileManager.h"
 #include "LogManager.h"
 #include "DataManager.h"
+
+// #include "hardware/MS5607.h"
+// #include "hardware/DFRobot_BMX160.h"
+// #include "hardware/PWM.h"
+// #include "hardware/SerialLinux.h"
+// #include "hardware/LED.h"
+
 
 #define DATA_FILENAME "data.csv"
 #define LOG_FILENAME "log.txt"
@@ -29,6 +37,7 @@ int main() {
     DataManager dataManager = DataManager(dataFile);
 
     /* Setup */
+    if (!fileManager.start()) return 1;
     if (!logManager.start()) return 1;
     if (!dataManager.start()) return 1;
