+++ launches/USLI_Launch_Beta/flight.h
@@ -1,0 +1,12 @@
+#ifndef _FLIGHT_H_
+#define _FLIGHT_H_
+
+#include <iostream>
+#include <fstream>
+#include <fcntl.h>
+#include <unistd.h>
+#include <time.h>
+
+int main();
+
+#endif
